// types.ts
export interface PostProfile {
  full_name: string | null;
  avatar_url: string | null;
}
// ...
export interface StoreFollower {
  user_id: string;
}

export interface StoreRating {
  user_id: string;
  rating: number;
}

export interface Store {
  id: string;
  created_at: string;
  name: string;
  description: string | null;
  user_id: string;
  image_url: string | null;
  profiles: PostProfile | null;
  store_followers: StoreFollower[];
  store_ratings: StoreRating[];
}